+++ docs/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2018-05-01 v1.1.0
+
+* Updated to Ubuntu 18.04.
+* Updated to docker-compose version 3.
+* Setup script enhancements.
+* Updated Ethminer.
+* Added option to control fan speed.
+
 ## 2018-02-11 v1.0.0
 
 * Initial version.